dashboard/requirements.txt
bidict==0.21.2
click==7.1.2
dnspython==1.16.0
eventlet==0.30.0
greenlet==2.0.1
gunicorn==20.0.4
h11==0.11.0
httptools==0.5.0
python-dotenv==0.15.0
python-engineio==4.0.0
python-socketio==5.0.4
PyYAML==6.0
six==1.15.0
uvicorn==0.13.2
uvloop==0.17.0
watchgod==0.6
websockets==10.4
wiringpi==2.60.1